Moved the ViewerMetadataWidget and its related widgets to a new subpackage for structure
