+++ src/napari_metadata/napari.yaml
@@ -6,7 +6,7 @@ contributions:
       python_name: napari_metadata.widgets:MetadataWidget
       title: Make layer metadata widget
     - id: napari-metadata.make_viewer_metadata_qwidget
-      python_name: napari_metadata.widgets:ViewerMetadataWidget
+      python_name: napari_metadata.viewer_widgets:ViewerMetadataWidget
       title: Make viewer metadata widget
   widgets:
     - command: napari-metadata.make_layer_metadata_qwidget
