+++ src/napari_metadata/viewer_widgets/__init__.py
@@ -1,0 +1,10 @@
+"""Viewer Widgets subpackage for napari-metadata.
+
+All Qt-dependent code lives here.
+"""
+
+from napari_metadata.viewer_widgets._viewer_metadata import (
+    ViewerMetadataWidget,
+)
+
+__all__ = ['ViewerMetadataWidget']
