+++ src/napari_metadata/widgets/__init__.py
@@ -4,6 +4,5 @@ All Qt-dependent code lives here.
 """
 
 from napari_metadata.widgets._main import MetadataWidget
-from napari_metadata.widgets._viewer_metadata import ViewerMetadataWidget
 
-__all__ = ['MetadataWidget', 'ViewerMetadataWidget']
+__all__ = ['MetadataWidget']
